main/requirements.txt.txt
numpy
opencv-python
ultralytics
pickle5
Pillow
tkintermapview
